requirements.txt
opencv-python>=4.8
numpy>=1.24
PyYAML>=6.0.1
rich>=13.7
torch>=2.0.0
torchvision>=0.15.0
onnx>=1.14.0
onnxruntime>=1.16.0
tensorrt>=8.6.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
Pillow>=10.0.0

# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-benchmark>=4.0.0
